+++ Mods ChoGGi/Solar Array Follows Sun/Code/Script.lua
@@ -2,10 +2,11 @@
 
 local function AddPanels(bld)
 	bld.ChoGGi_panels = {}
-  local attaches = bld:GetAttaches("SolarArrayPanel") or ""
-  for i = 1, #attaches do
-		bld.ChoGGi_panels[i] = attaches[i]
-  end
+	local c = 0
+	bld:ForEachAttach("SolarArrayPanel",function(a)
+		c = c + 1
+		bld.ChoGGi_panels[c] = a
+	end)
 end
 
 function OnMsg.BuildingInit(bld)